Add Battle View and Character/Enemy Selection
Added the view that will handle the battling portion of the program. Also allowed the user to choose different characters and go up against different enemies. Currently there are 2 options for enemies chosen at random with a 50/50 chance. Either the enemy is a "Default Enemy" that has randomly generated stats between 10 and 30, or the battle system can select another character from the database at random and use that, along with all of the associated stats.