Cleanup Code and Prepare for Battle Expansion
Reformatted a lot of code and changed the way the system handles battles. Now, there are dedicated sections to both player and enemy turns, and instead of a fight 'option' that was an 'int' data type, there is now both 'playerTurn' and 'enemyTurn' string being routed through the controller that will more clearly mark the player's choice and allow for the enemy to take different actions. The biggest current issue however, is the fact that the enemies do not take there turn.